Add files via upload
There are two files: fibonacci and poorlywrittenfibonacci.

The fibonacci.java will compile and run without errors. The poorlywrittenfibonacci, as the name implies, will not compile and run because of intentional errors. This is so we can use our code analysis and see if it will make changes it needs to run successfully.

The following errors are:
program will break if a non-integer input is given
loop should be i <= n, but it starts printing one more term than needed
missing semi-colon after input()